Move QueueUpdateStrategy and DownIdleTime time under Queues -> Slurm settings
